Updated tests. because removed GenesisConfig generic for oracle pallet
