+++ pallets/oracle/src/mock.rs
@@ -25,8 +25,10 @@ pub use primitives::{CurrencyId::Token, TokenSymbol::*};
 
 use crate::{
 	self as oracle,
-	dia::{DiaOracleAdapter, MockConvertPrice, MockOracleKeyConvertor, MockConvertMoment},
-	oracle_mock::{DataCollector, MockDiaOracle},
+	dia::DiaOracleAdapter,
+	oracle_mock::{
+		DataCollector, MockConvertMoment, MockConvertPrice, MockDiaOracle, MockOracleKeyConvertor,
+	},
 	Config, Error, OracleKeys,
 };
 
@@ -47,7 +49,7 @@ frame_support::construct_runtime!(
 
 		// Operational
 		Security: security::{Pallet, Call, Storage, Event<T>},
-		Oracle: oracle::{Pallet, Call, Config<T>, Storage, Event<T>},
+		Oracle: oracle::{Pallet, Call, Config, Storage, Event<T>},
 		Staking: staking::{Pallet, Storage, Event<T>},
 		Currency: currency::{Pallet},
 	}
@@ -205,11 +207,10 @@ impl ExtBuilder {
 	pub fn build() -> sp_io::TestExternalities {
 		let mut storage = frame_system::GenesisConfig::default().build_storage::<Test>().unwrap();
 
-		oracle::GenesisConfig {
-			max_delay: 0,
-			oracle_keys: vec![],
-		}
-		.assimilate_storage(&mut storage)
+		frame_support::traits::GenesisBuild::<Test>::assimilate_storage(
+			&oracle::GenesisConfig { oracle_keys: vec![], max_delay: 0 },
+			&mut storage,
+		)
 		.unwrap();
 
 		sp_io::TestExternalities::from(storage)
